+++ Actividad_Medio_Termino/Assets/Scripts/Framework/Controllers/Laser.cs
@@ -72,7 +72,6 @@ public class Laser : MonoBehaviour
                 enemyHealth.TakeDamage(1);
             }
         }
-        Debug.Log("Proyectil impactó en el objetivo: " + other.gameObject.name);
         // Destruir proyectil al colisionar
         Destroy(gameObject);
     }