+++ Actividad_Medio_Termino/Assets/Scripts/Framework/Controllers/Missile.cs
@@ -72,7 +72,6 @@ public class Missile : MonoBehaviour
                 enemyHealth.TakeDamage(10);
             }
         }
-        Debug.Log("Proyectil impactó en el objetivo: " + other.gameObject.name);
         // Destruir proyectil al colisionar
         Destroy(gameObject);
     }